Move all handlers to handler.go
